update migration city,weather . create model .schedule update  curl openweather api

+++ database/migrations/2020_08_20_171245_city.php
@@ -16,8 +16,8 @@ class City extends Migration
         Schema::create('city', function (Blueprint $table) {
             $table->id();
             $table->char('city_name', 16);
-            $table->float('lon');
-            $table->float('lat');
+            $table->double('lon');
+            $table->double('lat');
         });
     }
